Rearrange chart files
